refactor: 서버 라우터, 컨트롤러, 모델 모듈화
